pyproject.toml
[project]
name = "rl-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "plotly>=6.6.0",
]
